+++ Frontend/app/main/components/companies/companies.component.ts
@@ -5,22 +5,30 @@ import {Company} from '../../models/company.model';
 import {Observable} from 'rxjs/Observable';
 import {ToastsManager} from 'ng2-toastr';
 import {Router} from '@angular/router';
-import {CRMStore, CRMStoreService} from '../../services/crm-store.service';
+import {CRMStoreService} from '../../services/crm-store.service';
+import {slideTransitions} from '../../../shared/animations/transitions.animation';
 
 @Component({
 	selector: 'companies-component',
 	template: `
-		<button class="btn btn-block" [routerLink]="['/Add-Company']">Add A Company</button>
-		<table class="table table-bordered table-responsive table-hover">
-			<tbody>
-				<tr class="crm-list-item" *ngFor="let company of (companies$ | async)">
-					<td class="crm-list-item-title text-center" (click)="routeWithDispatch(company, ['/Company-Details'])">{{company.name}}</td>
-				</tr>
-			</tbody>
-		</table>
-	`
+	<button class="btn btn-block" [routerLink]="['/Add-Company']">Add A Company</button>
+	<table  class="table table-bordered table-responsive table-hover">
+		<tbody *ngIf="!!dataReady">
+			<tr class="crm-list-item" *ngFor="let company of (companies$ | async)">
+				<td  class="crm-list-item-title text-center" (click)="routeWithDispatch(company, ['/Company-Details'])">{{company.name}}</td>
+			</tr>
+		</tbody>
+		
+	</table>
+	`,
+	host: { '[@routeAnimation]': 'true' },
+	styles: [':host { display: block;}'],
+	animations: [
+	slideTransitions()
+]
 })
 export class CompaniesComponent implements OnInit {
+	public dataReady: boolean = false;
 	private companiesSource: BehaviorSubject<Company[]> = new BehaviorSubject<Company[]>([]);
 	public companies$: Observable<Company[]> = this.companiesSource.asObservable();
 	constructor(
@@ -36,7 +44,10 @@ export class CompaniesComponent implements OnInit {
 
 	private updateCompanies(): void {
 		this.crmService.getCompanies({})
-			.then((companies: Company[] )=> this.companiesSource.next(companies));
+			.then((companies: Company[]) => {
+				this.companiesSource.next(companies)
+					this.dataReady = true;
+			});
 
 	}