+++ Frontend/app/main/components/contacts/contacts.component.ts
@@ -8,23 +8,33 @@ import {Subscription} from 'rxjs/Subscription';
 import {Router} from '@angular/router';
 import {Company} from '../../models/company.model';
 import {ToastsManager} from 'ng2-toastr';
+import {slideTransitions} from '../../../shared/animations/transitions.animation';
 @Component({
 	selector:'contacts-component',
 	template: `
-		<button class="btn btn-block" [routerLink]="['/Add-Contact']">Add A Contact</button>
-		<table class="table table-bordered table-justified table-hover">
-			<tbody>
-				<tr class="crm-list-item" *ngFor="let contact of (contacts$ | async)">
-					<td class="crm-list-item-title" (click)="routeWithDispatch(contact, ['/Contact-Details'])">{{contact.name}}</td>
-					<td>
-						<span class="icon icon-share"(click)="routeWithDispatch(contact, ['/Quotes'])"></span>
-					</td>
-				</tr>
-			</tbody>
-		</table>
-	`
+		<div *ngIf="!!dataReady" >
+			<button class="btn btn-block" [routerLink]="['/Add-Contact']">Add A Contact</button>
+			<table class="table table-bordered table-justified table-hover">
+				<tbody>
+					<tr class="crm-list-item" *ngFor="let contact of (contacts$ | async)">
+						<td class="crm-list-item-title" (click)="routeWithDispatch(contact, ['/Contact-Details'])">{{contact.name}}</td>
+						<td>
+							<span class="icon icon-share" (click)="routeWithDispatch(contact, ['/Quotes'])"></span>
+						</td>
+					</tr>
+				</tbody>
+			</table>
+		</div>
+	`,
+	host: { '[@routeAnimation]': 'true' },
+	styles: [':host { display: block;}'],
+	animations: [
+		slideTransitions()
+	]
+
 })
 export class ContactsComponent implements OnInit, OnDestroy{
+	public dataReady: boolean = false;
 	private contactsSource: BehaviorSubject<Contact[]> = new BehaviorSubject<Contact[]>([]);
 	public contacts$: Observable<Contact[]> = this.contactsSource.asObservable();
 	public stateSub: Subscription;
@@ -51,11 +61,13 @@ export class ContactsComponent implements OnInit, OnDestroy{
 				this.crmData.getContacts({owner_id: state.selectedCompany.id})
 					.then((contacts: Contact[]) => {
 						this.contactsSource.next(contacts);
+						this.dataReady = true;
 					})
 			} else {
 				this.crmData.getContacts({})
 					.then((contacts: Contact[])=> {
 						this.contactsSource.next(contacts)
+						this.dataReady = true;
 					});
 			}
 		});