+++ Frontend/app/main/components/quotes/quote-details.component.ts
@@ -3,22 +3,24 @@ import {CRMDataService} from '../../services/crm-data.service';
 import {CRMStoreService} from '../../services/crm-store.service';
 import {Subscription} from 'rxjs/Subscription';
 import {BehaviorSubject} from 'rxjs/BehaviorSubject';
-import {newQuote, Quote, QuoteLine} from '../../models/quote.model';
+import {Quote, QuoteLine} from '../../models/quote.model';
 import {Observable} from 'rxjs/Observable';
 import {FormControl, FormGroup} from '@angular/forms';
 import * as _ from 'lodash';
 import {Router} from '@angular/router';
 import {ToastsManager} from 'ng2-toastr';
+import {slideTransitions} from '../../../shared/animations/transitions.animation';
 
 @Component({
 	selector: 'quote-details-component',
 	template: `
-	<h4>Quote Details</h4>
-	<hr>
+		<h4>Quote Details</h4>
+		<hr>
 		<quote-header [formGroup]="quoteForm">
 			<div class="row">
 				<h5>Quote Title: </h5>
-				<single-line-text-input-component *ngIf="!!quote.name" class="col-xs-12 text-center" [(model)]="quote.name"></single-line-text-input-component>
+				<single-line-text-input-component class="col-xs-12 text-center"
+												  [(model)]="quote.name"></single-line-text-input-component>
 				<hr/>
 			</div>
 		</quote-header>
@@ -27,19 +29,23 @@ import {ToastsManager} from 'ng2-toastr';
 			<h5>Quote Lines</h5>
 			<ul class="quote-list">
 				<hr/>
-				<li class="quote-list-line" *ngFor="let quoteLine of quoteLines; let i = index">
+				<li class="quote-list-line"
+					*ngFor="let quoteLine of (quoteLines$ | async); let i = index">
 					<span class="icon icon-cross pull-right" (click)="removeLine(i)"></span>
 					<div class="quote-list-line-title-wrapper">
-						<single-line-text-input-component label="Desc." [(model)]="quoteLine.desc"></single-line-text-input-component>
+						<single-line-text-input-component label="Desc."
+														  [(model)]="quoteLine.desc"></single-line-text-input-component>
 					</div>
 					<div class="quote-list-line-details">
 						<quote-line-inputs>
-							<single-line-text-input-component label="Units" [(model)]="quoteLine.unit"></single-line-text-input-component>
-							<single-line-text-input-component label="$" [(model)]="quoteLine.cost"></single-line-text-input-component>
+							<single-line-text-input-component label="Units"
+															  [(model)]="quoteLine.unit"></single-line-text-input-component>
+							<single-line-text-input-component label="$"
+															  [(model)]="quoteLine.cost"></single-line-text-input-component>
 						</quote-line-inputs>
 						<quote-line-options>
 							<span class="icon icon-arrow-up" (click)="onUp(i)"></span>
-							<span class="icon icon-arrow-down"(click)="onDown(i)"></span>
+							<span class="icon icon-arrow-down" (click)="onDown(i)"></span>
 						</quote-line-options>
 					</div>
 				</li>
@@ -49,29 +55,37 @@ import {ToastsManager} from 'ng2-toastr';
 		<quote-footer [formGroup]="newQuoteLineForm">
 			<div class="row">
 				<hr/>
-				<single-line-text-input-component label="Desc." [(model)]="newQuoteLine.desc"></single-line-text-input-component>
-				<single-line-text-input-component label="Unit Qty." [(model)]="newQuoteLine.unit"></single-line-text-input-component>
-				<single-line-text-input-component label="Cost" [(model)]="newQuoteLine.cost"></single-line-text-input-component>
+				<single-line-text-input-component label="Unit"
+												  [(model)]="newQuoteLine.unit"></single-line-text-input-component>
+				<single-line-text-input-component label="Desc."
+												  [(model)]="newQuoteLine.desc"></single-line-text-input-component>
+				<single-line-text-input-component label="Cost"
+												  [(model)]="newQuoteLine.cost"></single-line-text-input-component>
 				<button class="btn btn-lg" (click)="addLine()">Add Line</button>
 			</div>
 		</quote-footer>
 		<hr>
-	<div *ngIf="!checkRemove">
-		<button type="button" class="btn-success btn-lg pull-left" (click)="onSave()">Save</button>
-		<button type="button" class="btn-warning btn-lg pull-left" [routerLink]="['/Quotes']">Cancel</button>
-		<button type="button" class="btn-danger pull-right" (click)="onCheckRemove()">REMOVE</button>
-	</div>
-	<div class="check-remove" *ngIf="!!checkRemove">
-		<h4>Are you sure you want to remove {{quote.name}}?</h4>
-		<button type="button" class="btn-warning btn-lg pull-right" (click)="onCheckRemove()">Cancel</button>
-		<button type="button" class="btn-danger btn-lg pull-right" (click)="onRemove()">REMOVE</button>
-	</div>
-	`
+		<div *ngIf="!checkRemove">
+			<button type="button" class="btn-success btn-lg pull-left" (click)="onSave()">Save</button>
+			<button type="button" class="btn-warning btn-lg pull-left" [routerLink]="['/Quotes']">Cancel</button>
+			<button type="button" class="btn-danger pull-right" (click)="onCheckRemove()">REMOVE</button>
+		</div>
+		<div class="check-remove" *ngIf="!!checkRemove">
+			<h4>Are you sure you want to remove {{quote.name}}?</h4>
+			<button type="button" class="btn-warning btn-lg pull-right" (click)="onCheckRemove()">Cancel</button>
+			<button type="button" class="btn-danger btn-lg pull-right" (click)="onRemove()">REMOVE</button>
+		</div>
+	`,
+	host: { '[@routeAnimation]': 'true' },
+	styles: [':host { display: block;}'],
+	animations: [
+		slideTransitions()
+	]
 })
 export class QuoteDetailsComponent implements OnInit , OnDestroy{
 	public checkRemove: boolean = false;
 	private quoteLinesSource: BehaviorSubject<QuoteLine[]> = new BehaviorSubject<QuoteLine[]>([]);
-	public quoteLine$: Observable<QuoteLine[]> = this.quoteLinesSource.asObservable();
+	public quoteLines$: Observable<QuoteLine[]> = this.quoteLinesSource.asObservable();
 	public quoteLinesSub: Subscription;
 	public quoteLines: QuoteLine[] = [];
 	public newQuoteLine: QuoteLine = <QuoteLine>{};
@@ -108,19 +122,19 @@ export class QuoteDetailsComponent implements OnInit , OnDestroy{
 				this.crmData.getQuote({id: state.selectedQuote.id})
 					.then((quote: Quote) => {
 						this.quoteSource.next(quote);
-						this.quoteLinesSource.next(quote.quoteLines.sort((a, b) => {
+						const sorted = quote.quoteLines.sort((a, b) => {
 							if ( a.weight > b.weight ) {
 								return 1;
 							}
 							if (a.weight < b.weight) {
 								return -1;
 							}
-						})
-						);
+						});
+						this.quoteLinesSource.next(sorted);
 					})
 			}
 		});
-		this.quoteLinesSub = this.quoteLine$.subscribe(newLines => {
+		this.quoteLinesSub = this.quoteLines$.subscribe(newLines => {
 			this.quoteLines = newLines.sort((a, b) => {
 				if ( a.weight > b.weight ) {
 					return 1;